Fix bug where colliders' colors render strangely
when there's more than one textured obj

Cause: Forgot to set the bool that tells the shader that
no texture has been bound, so all the colliders were sampling
a non-existent texture. This results in undefined behaviour,
which is why the colliders rendered fine sometimes and then
broke when we started adding more textured objects.