Minor features: Toggle editor windows and debug drawing (#105)
* Update Editor

Added a filter system to filter different type of debug log such as
Error, Warning, Info, Events and normal text

Added a find text system that allows you to find based on input

* Update Editor

Added a command list mapped to command functions
Can type the commands in the command console
2 commands for now
ping which replies pong
test which opens test window if its closed

* update editor

Comment some code out that i'm not using still need to in the future
also fixed a bug

* Update Editor

Added 2 functions
public IsRenderingDebug - which returns whether the boolean for rendering debug is on
private ToggleDebugRender - which toggles the boolean for rendering

Added a button onto the test window that toggles the debug render
pressing F10 also toggles the debug render

* Add ability to toggle editor windows with ESC key

* Integrate Objects With Editor

Added creating objects using the object factory into the editor

* Integrate Object Components With Editor

Added the 3 components to have a drop down with editable variables

All of these are hardcoded for now, as in order to track variable, variable names we require reflection

Added
Private UpdateObjectList to update the list internally

* Update Component Editor

Integrated Renderer Component

Able to edit texture and colors

* Update Component Editor

Added Separators and Spaces to make it more readable
Added Mass for rigidbody because i missed it jn
Added alpha for the color on the renderer

* Update Component Editor

Made all the names of fields on the left instead of the right, as imgui window likes to cover the text from the right

Added a input value for rotation instead of only slider just for control (I may remove the slider completely)

* Add fix for clone function.

Missed legacy code phased out during rework of component pools

* Fix bug where colliders' colors render strangely

when there's more than one textured obj

Cause: Forgot to set the bool that tells the shader that
no texture has been bound, so all the colliders were sampling
a non-existent texture. This results in undefined behaviour,
which is why the colliders rendered fine sometimes and then
broke when we started adding more textured objects.

* Add base singleton class

Created a base singleton class as a template class with Myer's
Singleton pattern. This pattern avoids need for storing a pointer
to this instance of a class and other initializing issues.

Updated TimeManager and ResourceManager to inherit from this
Singleton class.

* Add fix for crashing at variant visit

SceneView optimization bugging out...

* Fix infinite loop for SceneView

* Fix crash when no entities

* Added logging to file

If the flag for is_error in add_log function is set to true, it will automatically flush the buffer and close the file. (otherwise it wont save, this is in expectation of an exception being thrown)

* Add ability to auto create Logs folder

* Add logging to ECS's throws

* Fix compiling issues

* Add ability to toggle debug shapes (on pressing F10)

* Update Singletons

Changed from unique pointer to inherit from singleton class
Removed eventhandler.cpp as all it was was a definition for the unique pointer

* Add comment

* Bug Fix

Fixed some merged errors

* Fix crash for illegal access

* Fix cloning crash

---------

Co-authored-by: Jtyz <jjjjtyz@gmail.com>
Co-authored-by: Foong Jun Wei <f.junwei@digipen.edu>
Co-authored-by: DonnHo <brandonhojj@gmail.com>