Fixed the memleak I think
the cause: I wasn't handling the behaviour of the dead rats properly. As such, because the death state
is triggered very abruptly, the logic of the rat's current state has a chance to run and attempt to change
the state of the rat again, causing a new state to get instantiated and lost in the void.